Break index.js into more components, move preload.js to common folder
