+++ Project Guru/src/components/Footer.js
@@ -1,0 +1,9 @@
+const Footer = () => {
+    return (
+        <footer>
+            <p>https://ephemeralxxv.github.io/ Copyright <span style = {{fontFamily: 'sans-serif'}}>&copy;</span> Project Guru, 2022. All rights reserved.</p>
+        </footer>
+    );
+}
+
+export default Footer;
