+++ Project Guru/src/components/Footer.module.css
@@ -1,0 +1,37 @@
+@media only screen and (max-width: 1280px) {
+    footer {
+        font-size: 80%;
+    }
+}
+
+@media only screen and (max-width: 992px) {
+    footer {
+        font-size: 70%;
+    }
+}
+
+@media only screen and (max-width: 600px) {
+    footer {
+        font-size: 60%;
+    }
+}
+
+@media only screen and (max-width: 420px) {
+    footer {
+        font-size: 50%;
+    }
+}
+
+footer {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    grid-row-start: 12;
+    grid-column: 1/11;
+    color: white;
+    background-color: black;
+}
+
+footer p {
+    margin: 0;
+}
