+++ Project Guru/src/pages/Index.module.css
@@ -23,15 +23,9 @@
     #page {
         grid-template-rows: 0.2fr repeat(10, minmax(0, 0.2fr)) 0.1fr;
     }
-    #menu_container {
+    #menuContainer {
         grid-template-columns: repeat(3, minmax(0, 1fr));
         grid-template-rows: minmax(0, 1fr) minmax(0, 1fr);
-    }
-}
-
-@media only screen and (max-width: 1280px) {
-    footer {
-        font-size: 80%;
     }
 }
 
@@ -39,12 +33,9 @@
     #page {
         grid-template-rows: 0.1fr repeat(10, minmax(0, 0.2fr)) 0.1fr;
     }
-    #menu_container {
+    #menuContainer {
         grid-template-columns: minmax(0, 1fr) minmax(0, 1fr);
         grid-template-rows: repeat(3, minmax(0, 1fr));
-    }
-    footer {
-        font-size: 70%;
     }
 }
 
@@ -52,58 +43,15 @@
     #page {
         grid-template-rows: 0.2fr 0.2fr repeat(8, minmax(0, 1fr)) 0.2fr 0.1fr;  /* Decreased gap between menu and navbar/footer */
     }
-    #menu_container {
+    #menuContainer {
         grid-template-columns: minmax(0, 1fr);
         grid-template-rows: repeat(6, minmax(0, 1fr));
     }
-    footer {
-        font-size: 60%;
-    }
-    
 }
 
-@media only screen and (max-width: 420px) {
-    footer {
-        font-size: 50%;
-    }
-}
-
-#navbarCollapsed {
-    cursor: pointer;
-    background: black;
-    grid-column: 1/12;
-    grid-row-end: 2;
-    opacity: 0.5;
-}
-
-#navbarExpanded {
-    position: sticky;
-    grid-column: 1/12;
-    grid-row: 2/6;
-    max-height: 0%;
-    overflow: hidden;
-    transition: max-height 0.2s ease-out;
-    background: white;
-}
-
-
-#menu_container {
+#menuContainer {
     overflow: inherit;
     grid-column: 2/10;
     grid-row: 3/11;
     display: grid;
-}
-
-footer {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    grid-row-start: 12;
-    grid-column: 1/11;
-    color: white;
-    background-color: black;
-}
-
-footer p {
-    margin: 0;
 }
